qa-system/requirements.txt
fastapi
uvicorn[standard]
openai
requests
streamlit
python-dotenv
pydantic
numpy
